feat(Movie): create data layer for the movies bounded context

+++ app/src/main/java/org/discovr/mobile/core/domain/Vote.kt
@@ -1,3 +1,3 @@
 package org.discovr.mobile.core.domain
 
-data class Vote(val average: Float, val count: Int)
+data class Vote(val average: Double, val count: Int)
